Use context-generic DelegateBuildCreateClientMessage
