+++ crates/cosmos/cosmos-chain-components/src/components/cosmos_to_cosmos.rs
@@ -1,9 +1,12 @@
 use cgp_core::prelude::*;
 use hermes_relayer_components::chain::impls::delegate::queries::client_state::QueryAndDecodeClientStateVia;
+use hermes_relayer_components::chain::traits::message_builders::create_client::CreateClientMessageBuilderComponent;
 use hermes_relayer_components::chain::traits::queries::client_state::{
     AllClientStatesQuerierComponent, ClientStateQuerierComponent,
 };
 use prost_types::Any;
+
+use crate::impls::client::create_client_message::BuildCosmosCreateClientMessage;
 
 pub struct CosmosToCosmosComponents;
 
@@ -14,5 +17,7 @@ delegate_components! {
             AllClientStatesQuerierComponent,
         ]:
             QueryAndDecodeClientStateVia<Any>,
+        CreateClientMessageBuilderComponent:
+            BuildCosmosCreateClientMessage,
     }
 }
