+++ crates/cosmos/cosmos-relayer/src/chain/impls/mod.rs
@@ -1,5 +1,4 @@
 pub mod connection_handshake_message;
-pub mod create_client_message;
 pub mod events;
 pub mod ext;
 pub mod fields;
