+++ crates/sovereign/sovereign-relayer/src/contexts/sovereign_chain.rs
@@ -1,18 +1,13 @@
 use cgp_core::prelude::*;
 use cgp_core::{delegate_all, ErrorRaiserComponent, ErrorTypeComponent};
 use cgp_error_eyre::{ProvideEyreError, RaiseDebugError};
-use hermes_cosmos_chain_components::components::delegate::DelegateCosmosChainComponents;
 use hermes_cosmos_relayer::contexts::chain::CosmosChain;
 use hermes_encoding_components::impls::default_encoding::GetDefaultEncoding;
 use hermes_encoding_components::traits::has_encoding::{
     DefaultEncodingGetterComponent, EncodingGetterComponent, EncodingTypeComponent, HasEncoding,
 };
-use hermes_relayer_components::chain::impls::delegate::queries::client_state::QueryAndDecodeClientStateVia;
 use hermes_relayer_components::chain::traits::payload_builders::connection_handshake::CanBuildConnectionHandshakePayloads;
 use hermes_relayer_components::chain::traits::payload_builders::update_client::CanBuildUpdateClientPayload;
-use hermes_relayer_components::chain::traits::queries::client_state::{
-    AllClientStatesBytesQuerierComponent, ClientStateQuerierComponent,
-};
 use hermes_relayer_components::chain::traits::types::client_state::HasClientStateType;
 use hermes_relayer_components::chain::traits::types::height::HasHeightType;
 use hermes_relayer_components::chain::traits::types::update_client::HasUpdateClientPayloadType;
@@ -28,7 +23,6 @@ use hermes_sovereign_chain_components::sovereign::traits::chain::data_chain::{
 };
 use hermes_sovereign_chain_components::sovereign::types::client_state::SovereignClientState;
 use hermes_sovereign_rollup_components::types::height::RollupHeight;
-use hermes_wasm_client_components::types::client_state::WasmClientState;
 
 use crate::contexts::encoding::{ProvideSovereignEncoding, SovereignEncoding};
 
@@ -80,24 +74,6 @@ delegate_components! {
         ]:
             ProvideSovereignEncoding,
         EncodingGetterComponent: GetDefaultEncoding,
-    }
-}
-
-pub struct SovereignCosmosComponents;
-
-delegate_components! {
-    SovereignCosmosComponents {
-        [
-            ClientStateQuerierComponent,
-            AllClientStatesBytesQuerierComponent,
-        ]:
-            QueryAndDecodeClientStateVia<WasmClientState>,
-    }
-}
-
-delegate_components! {
-    DelegateCosmosChainComponents {
-        SovereignChain: SovereignCosmosComponents,
     }
 }
